chnges 22april2021

+++ Modules/User/Http/Controllers/GoalItemController.php
@@ -5,6 +5,8 @@ use Modules\User\Http\Requests\GoalItemRequest;
 use Modules\User\Http\Services\GoalItemService;
 use Modules\User\Transformers\GoalItemTransformer;
 use Spatie\Fractal\Fractal;
+use Helper;
+use Illuminate\Http\Request;
 
 class GoalItemController extends Controller
 {
@@ -18,6 +20,7 @@ class GoalItemController extends Controller
     public function __construct(GoalItemService $service)
     {
         $this->service = $service;
+		$this->middleware('auth:api');
     }
 
     /**
@@ -26,24 +29,48 @@ class GoalItemController extends Controller
      *
      * @return Fractal
      */
-    public function store($user_id, GoalItemRequest $request): Fractal
+    public function store($user_id, Request $request): Fractal
     {
-        $goal_item = $this->service->store($user_id, $request->all());
 
+        try{
+            $user_id = Helper::customDecrypt($user_id);
+            $product_id = $request->product_id;
+            $product_id = Helper::customDecrypt($product_id);
+            $request['product_id'] = $product_id;
+            $rules = [
+                'product_id' => 'required|exists:products,id',
+            ];
 
-        return fractal($goal_item, new GoalItemTransformer());
+        
+            $validator = \Validator::make($request->all(), $rules);
+
+            if ($validator->fails())
+                return response()->json(['message' => 'The given data was invalid.', 'errors' => $validator->errors()], 422);
+
+            $goal_item = $this->service->store($user_id, $request->all());
+
+            return fractal($goal_item, new GoalItemTransformer());
+        }catch (\Throwable $th) {
+            return response()->json(['message' => 'Something get wrong! Please check user_id, product_id and try again.', 'errors' => $th->getMessage()], 402);
+        }
+
     }
-
+        
     /**
      * @param $user_id
      *
      * @return Fractal
      */
-    public function getUserGoalItem($user_id): Fractal
+    public function getUserGoalItem($user_id)
     {
-        $goal_item = $this->service->getUserGoalItem($user_id);
+        try{
+            $user_id = Helper::customDecrypt($user_id);
+            $goal_item = $this->service->getUserGoalItem($user_id);
 
-        return fractal($goal_item, new GoalItemTransformer());
+            return fractal($goal_item, new GoalItemTransformer());
+        }catch (\Throwable $th) {
+            return response()->json(['message' => 'Something get wrong! Please check user_id and try again.', 'errors' => $th->getMessage()], 402);
+        }
     }
 
    /**
@@ -52,12 +79,30 @@ class GoalItemController extends Controller
      *
      * @return Fractal
      */
-    public function removeGoalItem($user_id, GoalItemRequest $request): Fractal
+    public function removeGoalItem($user_id, Request $request): Fractal
     {
-        $goal_item = $this->service->remove($user_id, $request->all());
 
-        return fractal($goal_item, new GoalItemTransformer());
-        //return fractal($goal_item, new GoalItemTransformer());
+        try{
+            $user_id = Helper::customDecrypt($user_id);
+            $product_id = $request->product_id;
+            $product_id = Helper::customDecrypt($product_id);
+            $request['product_id'] = $product_id;
+            $rules = [
+                'product_id' => 'required|exists:products,id',
+            ];
+
+        
+            $validator = \Validator::make($request->all(), $rules);
+
+            if ($validator->fails())
+                return response()->json(['message' => 'The given data was invalid.', 'errors' => $validator->errors()], 422);
+
+            $goal_item = $this->service->remove($user_id, $request->all());
+
+            return fractal($goal_item, new GoalItemTransformer());
+        }catch (\Throwable $th) {
+            return response()->json(['message' => 'Something get wrong! Please check user_id, product_id and try again.', 'errors' => $th->getMessage()], 402);
+        }
     }
